# engine/graph-engine/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "KWeaver graph-engine and alg-server",
        "title": "kw_engine",
        "contact": {},
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/engine/v1/kg/": {
            "post": {
                "description": "graphql search",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Engine"
                ],
                "summary": "graphql search",
                "responses": {
                    "200": {
                        "description": "result data",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "500": {
                        "description": "EngineServer.ErrResourceErr: did not find configuration file ",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "utils.Error": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error": {},
                "errorCode": {
                    "type": "string"
                },
                "errorDetails": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                },
                "errorLink": {
                    "type": "string"
                },
                "solution": {
                    "type": "string"
                }
            }
        }
    }
}